requirements.txt
Flask==2.1.0
Werkzeug==2.0.1
gunicorn==20.1.0
python-dotenv==1.0.0
requests==2.31.0
